third_party/blink/web_tests/wpt_internal/document-transition/event-pseudo-name.html

<!DOCTYPE html>
<title>Shared transitions: event pseudo name</title>
<link rel="help" href="https://github.com/WICG/shared-element-transitions">
<link rel="author" href="mailto:vmpstr@chromium.org">
<link rel="match" href="web-animations-api-ref.html">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
:root { page-transition-tag: none; }
#first {
  background: blue;
  width: 100px;
  height: 100px;
  contain:  paint;
  page-transition-tag: shared;
}

html::page-transition-container(*),
html::page-transition-image-wrapper(*),
html::page-transition-incoming-image(*),
html::page-transition-outgoing-image(*) {
  animation-duration: 600s;
}

@keyframes fade-in {
  from { opacity: 0; }
}
html::page-transition-image-wrapper(*) {
  animation: fade-in 600s both;
}

</style>
<div id=first></div>
<script>
async_test(t => {
  let names = [];
  document.documentElement.addEventListener("animationstart", (e) => {
    names.push(e.pseudoElement);
    if (names.length == 4) {
      t.step(() => assert_true(names.includes("::page-transition-container(shared)")));
      t.step(() => assert_true(names.includes("::page-transition-image-wrapper(shared)")));
      t.step(() => assert_true(names.includes("::page-transition-incoming-image(shared)")));
      t.step(() => assert_true(names.includes("::page-transition-outgoing-image(shared)")));
      t.done();
    }
  });
  document.createDocumentTransition().start();
}, "verifies pseudo name includes a tag");

</script>
